+++ packages/react-native/src/NotifeeApiModule.ts
@@ -492,6 +492,8 @@ export default class NotifeeApiModule extends NotifeeNativeModule implements Mod
 
     const subscriber = this.emitter.addListener(
       kReactNativeNotifeeNotificationEvent,
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+      // @ts-ignore See https://github.com/facebook/react-native/pull/36462
       ({ type, detail }) => {
         observer({ type, detail });
       },